# sem11/src/spring-server-generated/src/main/java/io/swagger/model/User.java
package io.swagger.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.v3.oas.annotations.media.Schema;
import org.springframework.validation.annotation.Validated;
import javax.validation.Valid;
import javax.validation.constraints.*;

/**
 * User
 */
@Validated
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2023-08-12T11:28:20.035575131Z[GMT]")


public class User   {
  @JsonProperty("id")
  private Long id = null;

  @JsonProperty("name")
  private String name = null;

  @JsonProperty("email")
  private String email = null;

  @JsonProperty("passHash")
  private Long passHash = null;

  @JsonProperty("robotSNHash")
  private Long robotSNHash = null;

  public User id(Long id) {
    this.id = id;
    return this;
  }

  /**
   * Get id
   * @return id
   **/
  @Schema(example = "10", description = "")
  
    public Long getId() {
    return id;
  }

  public void setId(Long id) {
    this.id = id;
  }

  public User name(String name) {
    this.name = name;
    return this;
  }

  /**
   * Get name
   * @return name
   **/
  @Schema(example = "something", required = true, description = "")
      @NotNull

    public String getName() {
    return name;
  }

  public void setName(String name) {
    this.name = name;
  }

  public User email(String email) {
    this.email = email;
    return this;
  }

  /**
   * Get email
   * @return email
   **/
  @Schema(example = "somemail@mail.ru", required = true, description = "")
      @NotNull

    public String getEmail() {
    return email;
  }

  public void setEmail(String email) {
    this.email = email;
  }

  public User passHash(Long passHash) {
    this.passHash = passHash;
    return this;
  }

  /**
   * Get passHash
   * @return passHash
   **/
  @Schema(example = "23412", required = true, description = "")
      @NotNull

    public Long getPassHash() {
    return passHash;
  }

  public void setPassHash(Long passHash) {
    this.passHash = passHash;
  }

  public User robotSNHash(Long robotSNHash) {
    this.robotSNHash = robotSNHash;
    return this;
  }

  /**
   * Get robotSNHash
   * @return robotSNHash
   **/
  @Schema(example = "23412", required = true, description = "")
      @NotNull

    public Long getRobotSNHash() {
    return robotSNHash;
  }

  public void setRobotSNHash(Long robotSNHash) {
    this.robotSNHash = robotSNHash;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    User user = (User) o;
    return Objects.equals(this.id, user.id) &&
        Objects.equals(this.name, user.name) &&
        Objects.equals(this.email, user.email) &&
        Objects.equals(this.passHash, user.passHash) &&
        Objects.equals(this.robotSNHash, user.robotSNHash);
  }

  @Override
  public int hashCode() {
    return Objects.hash(id, name, email, passHash, robotSNHash);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class User {\n");
    
    sb.append("    id: ").append(toIndentedString(id)).append("\n");
    sb.append("    name: ").append(toIndentedString(name)).append("\n");
    sb.append("    email: ").append(toIndentedString(email)).append("\n");
    sb.append("    passHash: ").append(toIndentedString(passHash)).append("\n");
    sb.append("    robotSNHash: ").append(toIndentedString(robotSNHash)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}